requirements.txt
numpy==1.16.0
xlrd==1.2.0
PyMaxflow==1.2.0